Implemented first sync-test module
Copied and modifed Lizhuohua's sync_test module, left in the aspects of
the module that were simply just hello-world again. The original
implementation did not actually print the contents of the global spin
lock, so this implementation prints out those contents on removal.

I honestly do not know how Kbuild is interfacing right now with the
kernel build process, suffice to say currently this process *only* works
for compiling a helloworld.ko module (regardless of name) and I have
a vague understanding of why, but am too tired to figure this out right
now.