Added #[used] attribute to modinfo
Release builds of modules would remove generated modinfo strings without
the #[used] attribute. This update brings that in line with what would
be done in C. It still is not properly recognised though despite most
fields (save for author listing) appearing to be in the correct position
and order in the generated binaries.